+++ src/lib/file.h
@@ -42,7 +42,7 @@ public:
 	qint16 id() const;
 	QString name() const;
 	QString path() const;
-	qint32 size() const;
+	quint64 size() const;
 	Action action() const;
 	QString url() const;
 	QString userName() const;
@@ -50,7 +50,7 @@ public:
 	void setId(qint16);
 	void setName(QString);
 	void setPath(QString);
-	void setSize(qint32);
+	void setSize(quint64);
 	void setAction(Action);
 	void setUrl(QString);
 	void setUserName(QString);
@@ -59,7 +59,7 @@ protected:
 	qint16 m_id;
 	QString m_name;
 	QString m_path;
-	qint32 m_size;
+	quint64 m_size;
 	Action m_action;
 	QString m_url;
 	QString m_userName;
